// lib/types/culturalSpace.ts
export interface CulturalSpace {
  id: string;
  name: string;
  category?: string;
  address?: string;
  guname?: string;
  phone?: string;
  homepage?: string;
  openHours?: string;
  entranceFee?: string;
  closedDay?: string;
  seatCount?: number;
  description?: string;
  imageUrl?: string;
  lat?: number;
  lng?: number;
}